Adds editable preferences to utilities to decide install or not

+++ cookbooks/workstation/recipes/entertainment_utilities.rb
@@ -2,6 +2,9 @@
 # INTERNET
 # ============================================================
 
+is_google_chrome_install = node[:preferences][:internet][:google_chrome][:install]
+is_fast_install = node[:preferences][:internet][:fast][:install]
+
 # O Google Chrome é o navegador mais utilizado no Brasil e no
 # mundo atualmente, porém, só está disponível para distribuições
 # Linux de 64 bits.
@@ -12,22 +15,32 @@ apt_repository 'Add: Google Chrome' do
   distribution 'stable'
   components %w(main)
   action :add
+
+  only_if is_google_chrome_install
 end
 
 package 'Install: Google Chrome' do
   package_name 'google-chrome-stable'
   options '--force-yes'
   action :install
+
+  only_if is_google_chrome_install
 end
 
 # Test your internet download speed from terminal
 execute 'Install: fast' do
   command 'sudo snap install fast'
+  only_if is_fast_install
 end
 
 # ============================================================
 # MESSENGERS AND SOCIAL
 # ============================================================
+
+is_install_telegram = node[:preferences][:messengers_social][:telegram][:install]
+is_install_skype = node[:preferences][:messengers_social][:skype][:install]
+is_install_slack = node[:preferences][:messengers_social][:slack][:install]
+
 # Caution: Slack for Linux is in beta. We’re still busy adding
 # features and ironing out potential issues.
 # 
@@ -41,6 +54,7 @@ end
 # desk or on the go.
 execute 'Install: Slack' do
   command 'sudo snap install slack --classic'
+  only_if is_install_slack
 end
 
 # Skype keeps the world talking. Say “hello” with an instant
@@ -48,6 +62,7 @@ end
 # on phones, tablets, PCs, Macs and Linux.
 execute 'Install: Skype' do
   command 'sudo snap install skype --classic'
+  only_if is_install_skype
 end
 
 # Pure instant messaging — simple, fast, secure, and synced across
@@ -55,12 +70,20 @@ end
 # half years.
 execute 'Install: Telegram' do
   command 'sudo snap install telegram-desktop'
+  only_if is_install_telegram
 end
 
 
 # ============================================================
 # OFFICE
 # ============================================================
+
+is_remove_libre_office = node[:preferences][:office][:libre_office][:install]
+
+is_install_gmail_desktop = node[:preferences][:office][:gmail_desktop][:install]
+is_install_only_office = node[:preferences][:office][:only_office][:install]
+is_install_csbooks = node[:preferences][:office][:csbooks][:install]
+is_install_buka = node[:preferences][:office][:buka][:install]
 
 # ONLYOFFICE Desktop Editors is a free open source office suite
 # that combines text, spreadsheet and presentation editors
@@ -74,10 +97,12 @@ execute 'Removes: libreoffice' do
     sudo apt-get clean
     sudo apt-get autoremove
   EOH
+  only_if is_remove_libre_office
 end
 
 execute 'Install: only-office' do
   command 'sudo snap install onlyoffice-desktopeditors'
+  only_if is_install_only_office
 end
 
 # Gmail Desktop is a cross-platform dedicated Gmail app written in
@@ -85,11 +110,13 @@ end
 # default mailto handling, and more!
 execute 'Install: Gmail Desktop' do
   command 'sudo snap install gmail-desktop'
+  only_if is_install_gmail_desktop
 end
 
 # EBook Management
 execute 'Install: Buka' do
   command 'sudo snap install buka'
+  only_if is_install_buka
 end
 
 # csBooks is a smart solution to manage all your PDF, EPUB and
@@ -101,12 +128,16 @@ end
 # read status for them.
 execute 'Install: csBooks' do
   command 'sudo snap install csbooks'
+  only_if is_install_csbooks
 end
 
 
 # ============================================================
 # PERSONALISATION
 # ============================================================
+
+is_install_disk_space_saver = node[:preferences][:personalisation][:disk_space_saver][:install]
+
 # Disk Space Saver is a disk space analyzer, that will help you
 # to find large space hogs on hard drive in seconds! And free
 # up disk space easily in a click!
@@ -118,14 +149,22 @@ end
 # & delete large files and folders in a simple click.
 execute 'Install: Disk Space Saver' do
   command 'sudo snap install disk-space-saver'
+  only_if is_install_disk_space_saver
 end
 
 
 # ============================================================
 # EDUCATION
 # ============================================================
+is_install_kmplot = node[:preferences][:education][:kmplot][:install]
+is_install_step = node[:preferences][:education][:step][:install]
+is_install_truthtables = node[:preferences][:education][:truthtables][:install]
+is_install_atomify = node[:preferences][:education][:atomify][:install]
+is_install_qalculate = node[:preferences][:education][:qalculate][:install]
+
 execute 'Install: kmplot' do
   command 'sudo snap install kmplot'
+  only_if is_install_kmplot
 end
 
 # Step is an interactive physical simulator. It allows you to
@@ -139,11 +178,13 @@ end
 # learn but feel how physics works!
 execute 'Install: step' do
   command 'sudo snap install step'
+  only_if is_install_step
 end
 
 # Generates truth tables. Allows 30 different variables.
 execute 'Install: truthtables' do
   command 'sudo snap install truthtables'
+  only_if is_install_truthtables
 end
 
 # Atomify LAMMPS is an easy-to-use visualizer and editor for the
@@ -155,6 +196,7 @@ end
 # is open source software written in C++ built on top of Qt.
 execute 'Install: atomify' do
   command 'sudo snap install atomify'
+  only_if is_install_atomify
 end
 
 # Qalculate! is a multi-purpose cross-platform desktop calculator.
@@ -166,8 +208,9 @@ end
 # symbolic calculations (including integrals and equations),
 # arbitrary precision, uncertainty propagation, interval arithmetic,
 # plotting, and a user-friendly interface.
-execute 'Install: QCalculate' do
+execute 'Install: qalculate' do
   command 'sudo snap install qalculate'
+  only_if is_install_qalculate
 end
 
 
@@ -175,23 +218,32 @@ end
 # PHOTO AND VIDEO
 # ============================================================
 
+is_install_inkscape = node[:preferences][:photo_video][:inkscape][:install]
+is_install_photoscape = node[:preferences][:photo_video][:photoscape][:install]
+is_install_electron_player = node[:preferences][:photo_video][:electron_player][:install]
+is_install_youtube_dl = node[:preferences][:photo_video][:youtube_dl][:install]
+is_install_gimp = node[:preferences][:photo_video][:gimp][:install]
+
 # An Open Source vector graphics editor, with capabilities similar
 # to Illustrator, CorelDraw, or Xara X, using the W3C standard
 # Scalable Vector Graphics (SVG) file format.
 execute 'Install: inkscape' do
   command 'sudo snap install inkscape'
+  only_if is_install_inkscape
 end
 
 # PhotoScape is a fun and easy photo editing software that enables
 # you to fix and enhance photos.
 execute 'Install: photoscape' do
   command 'sudo snap install photoscape'
+  only_if is_install_photoscape
 end
 
 # An Electron Based Web Video Services Player. Supporting Netflix,
 # Youtube, Twitch, Floatplane And More.
 execute 'Install: electron-player' do
   command 'sudo snap install electronplayer'
+  only_if is_install_electron_player
 end
 
 # Is a command-line program to download videos from YouTube.com and
@@ -202,6 +254,7 @@ end
 # use it however you like.
 execute 'Install: youtube-dl' do
   command 'sudo snap install youtube-dl'
+  only_if is_install_youtube_dl
 end
 
 # Whether you are a graphic designer, photographer, illustrator, or
@@ -210,11 +263,18 @@ end
 # thanks to many customization options and 3rd party plugins.
 execute 'Install: gimp' do
   command 'sudo snap install gimp'
+  only_if is_install_gimp
 end
 
 # ============================================================
 # MUSIC AND AUDIO
 # ============================================================
+
+is_install_spotify = node[:preferences][:music_audio][:spotify][:install]
+is_install_audacity = node[:preferences][:music_audio][:audacity][:install]
+is_install_musescore = node[:preferences][:music_audio][:musescore][:install]
+is_install_tuxguitar = node[:preferences][:music_audio][:tuxguitar][:install]
+is_install_deep_voice_recorder = node[:preferences][:music_audio][:deep_voice_recorder][:install]
 
 # Stream the tracks you love instantly, browse the charts or
 # fire up readymade playlists in every genre and mood. Radio
@@ -223,6 +283,7 @@ end
 # just for you.
 execute 'Install: spotify' do
   command 'sudo snap install spotify'
+  only_if is_install_spotify
 end
 
 # Audacity is a free, easy-to-use, multi-track audio editor and
@@ -230,6 +291,7 @@ end
 # systems. The interface is translated into many languages.
 execute 'Install: audacity' do
   command 'sudo snap install audacity'
+  only_if is_install_audacity
 end
 
 # MuseScore is the world's leading free and open-source software
@@ -238,6 +300,7 @@ end
 # Linux.
 execute 'Install: musescore' do
   command 'sudo snap install musescore'
+  only_if is_install_musescore
 end
 
 # TuxGuitar is a free, open source tablature editor, which includes
@@ -245,10 +308,12 @@ end
 # export of Guitar Pro gp3, gp4, and gp5 files
 execute 'Install: tuxguitar' do
   command 'sudo snap install tuxguitar-vs'
+  only_if is_install_tuxguitar
 end
 
 # This is deepin-voice-recorder snap application. It can be used to
 # record and play voice files.
 execute 'Install: deep-voice-recorder' do
   command 'sudo snap install deepin-voice-recorder'
+  only_if is_install_deep_voice_recorder
 end
